requirements.txt
python-chess==1.10.0
zstandard==0.22.0
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
h5py==3.10.0
pyyaml==6.0.1
tqdm==4.66.4
matplotlib==3.8.4
pytest==8.1.1
